backend/requirements.txt
# Core dependencies
flask==3.0.*
functions-framework==3.5.*
google-cloud-firestore==2.13.*
google-auth==2.27.*
google-auth-oauthlib==1.2.*
google-auth-httplib2==0.1.*
google-api-python-client==2.116.*
pyjwt==2.8.*
requests==2.32.*
python-dateutil==2.8.*
fastapi==0.104.1
pydantic==2.4.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Monitoring
opencensus-ext-stackdriver==0.8.0
opencensus==0.11.0